Fix the robot configuration for HW hand
After the merge from joelryt SOP-robot fork there was issue that
HW hand is not working. This fixes the robot configuration
so that you are able to start the HW hand succesfully.
On the other hand this leaves the functionality of the head out,
because of change to dynamixel.yaml.

Fake hardware and sensor variables set to false in
inmoov.urdf.xacro to be able to start the hand without starting
the fake hardware.

Servo IDs that are not found are commented out in dynamixel.yaml.